requirements.txt
lightgbm==4.6.0
numpy==1.22.4
numpy==1.20.3
optuna==4.2.0
rank_bm25==0.2.2
scikit_learn==0.24.2
scipy==1.7.1
tensorflow_gpu==2.8.0
tqdm==4.62.3
transformers==4.36.0
transformers==4.16.2
